requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
Flask==0.10.1
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
Werkzeug==0.9.6
dj-database-url==0.3.0
django-filter==0.7
django-taggit==0.12
django-toolbox==0.1
djangorestframework==2.3.13
docutils==0.11
gunicorn==18.0
itsdangerous==0.24
nose==1.3.3
psycopg2==2.5.3
pylast==0.5.11
